implement most of GA for Similarity Measure, and using IFPUG to test

+++ src/EvaluationMethod/EvaluateMRE.java
@@ -57,7 +57,7 @@ public class EvaluateMRE {
 			}
 		}
 		
-		Pred = num/size;
+		Pred = (double)num/size;
 		return Pred;
 	}
